==> notebooks/5_Apache_Spark.ipynb <==
from pyspark.sql import SparkSession, Row
from pyspark.sql.functions import col, avg, when, round

# Inicijalizacija Spark sesije
spark = SparkSession.builder.appName("CovidAnalysisSpark").getOrCreate()

#Učitavanje detaset-a 
df_spark = spark.read.csv('../data/Covid_Data_Clean.csv', header=True, inferSchema=True)

#Prikaz prvih 5 redova dataset-a
df_spark.show(5)

#Prikaz tipova podataka
df_spark.printSchema()


# ISTRAŽIVAČKO PITANJE 1 - Distribucije ishoda virusa prema starosti i polu

mortality_by_age = df_spark.groupBy("AGE_GROUP") \
    .agg(round(avg("DEATH")*100, 2).alias("death_rate_percent")) \
    .orderBy("AGE_GROUP")
mortality_by_age.show()

Na osnovu dobijene tabele možemo zaključiti da procenat smrtnosti znatno raste sa povećanjem broja godina. Primećuje se da je stopa smrtnosti za uzrast od 0-19 godina 1,57%, dok je za 80+ čak skoro 53%.

#grupisanje smrti po polu
mortality_by_sex = df_spark.groupBy("SEX") \
    .agg(round(avg("DEATH")*100, 2).alias("death_rate_percent"))
mortality_by_sex.show()

Na osnovu dobijene tabele možemo zaključiti da muškarci (0) imaju malo veću stopu smrtnosti - 16% naprema 10% za žene

# ISTRAŽIVAČKO PITANJE 2 - Faktori hospitalizacije

# Uticaj gojaznosti na hospitalizaciju
hospital_by_obesity = df_spark.groupBy("OBESITY") \
    .agg(round(avg("HOSPITALIZED")*100, 2).alias("hospital_rate_percent"))
hospital_by_obesity.show()

Možemo zaključiti da gojazni pacijenti imaju veću šansu za hospitalizaciju, skoro 35%, naprema pacijenata koji nisu gojazni -oko 26%

# Procenat hospitalizacije za pacijente sa dijabetesom
hospital_by_diabetes = df_spark.groupBy("DIABETES") \
    .agg(round(avg("HOSPITALIZED")*100, 2).alias("hospital_rate_percent"))
hospital_by_diabetes.show()

Možemo videti da je znatno veća verovatnoća hospitalizacije pacijenata sa dijabetesom - oko 55% 

#prikaz korelacije
corr_diabetes = df_spark.stat.corr("DIABETES", "HOSPITALIZED")
corr_obesity = df_spark.stat.corr("OBESITY", "HOSPITALIZED")

print("Factor       Correlation")
print("------------------------")
print(f"DIABETES     {corr_diabetes}")
print(f"OBESITY      {corr_obesity}")

Rezultati pokazuju pozitivnu korelaciju (oko 0.27) između dijabetesa i hospitalizacije, što ukazuje da pacijenti sa dijabetesom imaju povećanu verovatnoću hospitalizacije.
Korelacija između gojaznosti i hospitalizacije je veoma slaba (oko 0.07), što sugeriše da gojaznost nije snažan prediktor hospitalizacije.

# ISTRAŽIVAČKO PITANJE 3 - Komorbiditeti i kritična stanja

#verovatnoća da pacijenti idu na ICU u odnosu na broj hroničnih bolesti
df_spark.groupBy("COMORBIDITY_GROUP").agg(round(avg("ICU")*100,2).alias("icu_prob")).orderBy("COMORBIDITY_GROUP").show()

Vidimo da što je veći broj bolesti, to je veća verovatnoća odlaska na intenzivnu negu

# Filtriranje pacijenata sa 3+ hroničnih bolesti
high_risk_patients = df_spark.filter(col("COMORBIDITY_GROUP") == "3+")

# Grupisanje po intubaciji i brojanje
intubed_count = high_risk_patients.groupBy("INTUBED").count()
intubed_count.show()

Možemo videti da ima 7629 pacijenata koji imaju 3 ili više bolesti i koji su završili na respiratoru

# ISTRAŽIVAČKO PITANJE 4: Pneumonija, starost i smrtnost

# Filter pacijenata sa pneumonijom i grupisanje po AGE_GROUP
mortality_with_pneumonia = df_spark.filter(col("PNEUMONIA") == 1) \
    .groupBy("AGE_GROUP") \
    .agg(round(avg("DEATH")*100,2).alias("death_rate")) \
    .orderBy("AGE_GROUP")

mortality_with_pneumonia.show()

Što je starija osoba koja ima preumiju, veća je šansa smrtnosti - za osobe sa preko 80 godina čak 72,4% 

# Grupisanje po TOBACCO i računanje prosečne pneumonije
pneumonia_by_smoking = df_spark.groupBy("TOBACCO") \
    .agg(round(avg(col("PNEUMONIA"))*100,2).alias("pneumonia_rate"))

pneumonia_by_smoking.show()

Zabeležena je veća stopa preumonije kod pušača (24,2%), dok je kod nepušača rasprostranjena u 21,52% slučaja
